+++ packages/core/src/base/custom_event.ts
@@ -13,7 +13,7 @@
 // You should have received a copy of the GNU Affero General Public License
 // along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
-import { NoReactiveSymbol } from "../runtime/reactive/reactive";
+import { NoReactiveSymbol } from "../runtime/reactive";
 
 class CustomEvent<T = unknown> {
   [NoReactiveSymbol] = true;
